don't share show_diffs code
